Move business logic out of TcUserDetails and into TcUserDetailsService
