+++ server/src/main/java/org/tctalent/server/security/TcUserDetails.java
@@ -16,7 +16,6 @@
 
 package org.tctalent.server.security;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
@@ -24,15 +23,12 @@ import java.util.stream.Collectors;
 import org.springframework.lang.NonNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.tctalent.server.model.db.Role;
 import org.tctalent.server.model.db.User;
 
 /**
- * Talent Catalog's implementation of Spring's {@link UserDetails}, retrieving user data from database, setting
- * authorities based on the user's {@link Role} and exposing the corresponding underlying
- * {@link User} object through {@link #getUser()}.
+ * Talent Catalog's implementation of Spring's {@link UserDetails}, exposing the corresponding
+ * underlying {@link User} object through {@link #getUser()}.
  * <p/>
  * Note that this is what is returned by {@link Authentication#getPrincipal()}
  */
@@ -40,19 +36,11 @@ public class TcUserDetails implements UserDetails {
     private static final String ROLE_PREFIX = "ROLE_"; // Spring default
 
     private User user;
-    private final List<GrantedAuthority> authorities;
-    private final Set<String> authorityStrings;
+    private List<GrantedAuthority> authorities;
+    private Set<String> authorityStrings;
 
     public TcUserDetails(@NonNull User user) {
         this.user = user;
-
-        //For now, we hard code the authorities based on the user.
-        authorities = createAuthorities(user);
-
-        //Construct Set of Strings to efficiently support hasAuthority(String) and hasRole(String)
-        this.authorityStrings = authorities.stream()
-            .map(GrantedAuthority::getAuthority)
-            .collect(Collectors.toUnmodifiableSet());
     }
 
     public boolean hasAuthority(String authority) {
@@ -79,45 +67,6 @@ public class TcUserDetails implements UserDetails {
         return false;
     }
 
-    /**
-     * In future, we may store user permissions in a separate db table and retrieve them here.
-     * @param user User to create authorities for
-     * @return List of authorities
-     */
-    private List<GrantedAuthority> createAuthorities(@NonNull User user) {
-        List<GrantedAuthority> authorities = new ArrayList<>();
-
-        // If read only is checked assign the read only role
-        if(user.getReadOnly()){
-            authorities.add(new SimpleGrantedAuthority("ROLE_READONLY"));
-        } else if (user.getRole().equals(Role.systemadmin)) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_SYSTEMADMIN"));
-        } else if (user.getRole().equals(Role.admin)) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_ADMIN"));
-        } else if (user.getRole().equals(Role.partneradmin)) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_PARTNERADMIN"));
-        } else if (user.getRole().equals(Role.semilimited)) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_SEMILIMITED"));
-        } else if (user.getRole().equals(Role.limited)) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_LIMITED"));
-        } else {
-            authorities.add(new SimpleGrantedAuthority("ROLE_USER"));
-        }
-
-        //For now we hard code some particular test users as our chat test users
-        if("TestChattingCandidate".equals(user.getUsername())){
-            authorities.add(new SimpleGrantedAuthority("CHAT_SUBSCRIBE"));
-        }
-        if("TestChattingJobCreator".equals(user.getUsername())){
-            authorities.add(new SimpleGrantedAuthority("CHAT_SUBSCRIBE"));
-        }
-        if("TestChattingSourcePartner".equals(user.getUsername())){
-            authorities.add(new SimpleGrantedAuthority("CHAT_SUBSCRIBE"));
-        }
-
-        return authorities;
-    }
-
     public @NonNull User getUser() {
         return user;
     }
@@ -129,6 +78,15 @@ public class TcUserDetails implements UserDetails {
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return authorities;
+    }
+
+    public void setAuthorities(List<GrantedAuthority> authorities) {
+        this.authorities = authorities;
+
+        //Construct Set of Strings to efficiently support hasAuthority(String) and hasRole(String)
+        this.authorityStrings = authorities.stream()
+            .map(GrantedAuthority::getAuthority)
+            .collect(Collectors.toUnmodifiableSet());
     }
 
     @Override
