+++ server/src/test/java/org/tctalent/server/security/TcUserDetailsTest.java
@@ -18,18 +18,15 @@ package org.tctalent.server.security;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import java.util.Collection;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.security.core.GrantedAuthority;
 import org.tctalent.server.model.db.Role;
 import org.tctalent.server.model.db.User;
 
@@ -53,93 +50,6 @@ class TcUserDetailsTest {
   @AfterEach
   void tearDown() throws Exception {
     mocks.close();
-  }
-
-  @Test
-  void constructor_withReadOnly_setsReadOnlyAuthority() {
-    when(user.getReadOnly()).thenReturn(true);
-    when(user.getRole()).thenReturn(Role.admin); // Role should be ignored if read-only
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_READONLY")), "Authority should be ROLE_READONLY");
-    assertTrue(tcUserDetails.hasRole("READONLY"), "Authority should be ROLE_READONLY");
-    assertTrue(tcUserDetails.hasAnyRole("X", "READONLY", "Y"), "Authority should be ROLE_READONLY");
-
-  }
-
-  @Test
-  void constructor_withSystemAdminRole_setsSystemAdminAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.systemadmin);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_SYSTEMADMIN")), "Authority should be ROLE_SYSTEMADMIN");
-  }
-
-  @Test
-  void constructor_withAdminRole_setsAdminAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.admin);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")), "Authority should be ROLE_ADMIN");
-  }
-
-  @Test
-  void constructor_withPartnerAdminRole_setsPartnerAdminAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.partneradmin);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_PARTNERADMIN")), "Authority should be ROLE_PARTNERADMIN");
-  }
-
-  @Test
-  void constructor_withSemiLimitedRole_setsSemiLimitedAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.semilimited);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_SEMILIMITED")), "Authority should be ROLE_SEMILIMITED");
-  }
-
-  @Test
-  void constructor_withLimitedRole_setsLimitedAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.limited);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_LIMITED")), "Authority should be ROLE_LIMITED");
-  }
-
-  @Test
-  void constructor_withUserRole_setsUserAuthority() {
-    when(user.getReadOnly()).thenReturn(false);
-    when(user.getRole()).thenReturn(Role.user);
-
-    tcUserDetails = new TcUserDetails(user);
-
-    Collection<? extends GrantedAuthority> authorities = tcUserDetails.getAuthorities();
-    assertEquals(1, authorities.size(), "Should have exactly one authority");
-    assertTrue(authorities.stream().anyMatch(a -> a.getAuthority().equals("ROLE_USER")), "Authority should be ROLE_USER");
   }
 
   @Test
@@ -207,11 +117,5 @@ class TcUserDetailsTest {
     tcUserDetails = new TcUserDetails(user);
 
     assertTrue(tcUserDetails.isEnabled(), "isEnabled should always return true");
-  }
-
-  @Test
-  void constructor_withNullUser_throwsNullPointerException() {
-    assertThrows(NullPointerException.class, () -> new TcUserDetails(null),
-        "Constructor should throw NullPointerException for null User due to @NotNull");
   }
 }
